feat: preserve toml class + track table if subsetting returns a list
When subsetting a toml object returns a list the toml class is now preserved. A
new attribute, "table", is added to track the current location in the nested
table. "table" is `NULL` when the toml object is at the root of the file it was
parsed from. "table" accumulates across repeated subsetting operations.

A simple example of a nested table is in `inst/toml/ex_nested_tables.toml`